Refactor: Delete SceneDelegate, add constraint to searchField
